gh-653: (refactor) seperate concerns and clean up get_meta pkg
There were namely these issues that made working with and extending the
current package problematic:

1.) a lot of concerns were mixed up in a single file that made it
incredibly difficult to find what you were looking for. This single file
has now been split up into 9 files

2.) The code itself used unnecessarily long function and variable names
all prefixed with the name of the package, etc. This redundant info made
the code very hard to read. This has been simplified.

3.) There was some obvious low-hanging fruit regarding repition in the
big switch-case to split up by property type. This has been extracted
into a function and made much more concise.

Overall I feel like the package is now in a decent state to start
building new features on top of the exisitng code base.
