--- graphqlapi/local/get_meta/dynamic_generation_getmeta.go
@@ -1,721 +1,0 @@
-/*                          _       _
- *__      _____  __ ___   ___  __ _| |_ ___
- *\ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
- * \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
- *  \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
- *
- * Copyright © 2016 - 2018 Weaviate. All rights reserved.
- * LICENSE: https://github.com/creativesoftwarefdn/weaviate/blob/develop/LICENSE.md
- * AUTHOR: Bob van Luijt (bob@kub.design)
- * See www.creativesoftwarefdn.org for details
- * Contact: @CreativeSofwFdn / bob@kub.design
- */
-
-// Package Local provides the Local graphql endpoint for Weaviate
-package get_meta
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/creativesoftwarefdn/weaviate/database/schema"
-	"github.com/creativesoftwarefdn/weaviate/graphqlapi/descriptions"
-	"github.com/creativesoftwarefdn/weaviate/graphqlapi/local/common_filters"
-	"github.com/creativesoftwarefdn/weaviate/models"
-	"github.com/graphql-go/graphql"
-)
-
-// Build the dynamically generated GetMeta Things part of the schema
-func genLocalMetaClassFieldsFromSchema(databaseSchema []*models.SemanticSchemaClass, classParentTypeIsAction bool) (*graphql.Object, error) {
-	classFields := graphql.Fields{}
-	kindName := "Things"
-	name := "WeaviateLocalGetMetaThingsObj"
-	description := descriptions.LocalGetMetaThingsObjDesc
-	if classParentTypeIsAction {
-		name = "WeaviateLocalGetMetaActionsObj"
-		description = descriptions.LocalGetMetaActionsObjDesc
-		kindName = "Actions"
-	}
-
-	for _, class := range databaseSchema {
-		field, err := genSingleLocalMetaClassField(kindName, class, class.Description)
-
-		if err != nil {
-			return nil, err
-		}
-
-		classFields[class.Class] = field
-	}
-
-	LocalGetMetaClasses := graphql.ObjectConfig{
-		Name:        name,
-		Fields:      classFields,
-		Description: description,
-	}
-
-	return graphql.NewObject(LocalGetMetaClasses), nil
-}
-
-func genSingleLocalMetaClassField(kindName string, class *models.SemanticSchemaClass, description string) (*graphql.Field, error) {
-	metaClassName := fmt.Sprintf("Meta%s", class.Class)
-
-	singleClassPropertyFields := graphql.ObjectConfig{
-		Name: metaClassName,
-		Fields: (graphql.FieldsThunk)(func() graphql.Fields {
-			singleClassPropertyFields, err := genSingleLocalMetaClassPropertyFields(class)
-
-			if err != nil {
-				panic("Failed to assemble single Local Meta Class field")
-			}
-
-			return singleClassPropertyFields
-		}),
-		Description: description,
-	}
-
-	singleClassPropertyFieldsObject := graphql.NewObject(singleClassPropertyFields)
-	singleClassPropertyFieldsField := &graphql.Field{
-		Type:        singleClassPropertyFieldsObject,
-		Description: description,
-		Args: graphql.FieldConfigArgument{
-			"first": &graphql.ArgumentConfig{
-				Description: descriptions.FirstDesc,
-				Type:        graphql.Int,
-			},
-			"after": &graphql.ArgumentConfig{
-				Description: descriptions.AfterDesc,
-				Type:        graphql.Int,
-			},
-			"where": &graphql.ArgumentConfig{
-				Description: descriptions.LocalGetWhereDesc,
-				Type: graphql.NewInputObject(
-					graphql.InputObjectConfig{
-						Name:        fmt.Sprintf("WeaviateLocalGetMeta%s%sWhereInpObj", kindName, class.Class),
-						Fields:      common_filters.BuildNew(fmt.Sprintf("WeaviateLocalGetMeta%s%s", kindName, class.Class)),
-						Description: descriptions.LocalGetWhereInpObjDesc,
-					},
-				),
-			},
-		},
-		Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-			source, ok := p.Source.(map[string]interface{})
-			if !ok {
-				return nil, fmt.Errorf("expected source to be a map, but was %t", p.Source)
-			}
-
-			resolver, ok := source["Resolver"].(Resolver)
-			if !ok {
-				return nil, fmt.Errorf("expected source to contain a usable Resolver, but was %t", p.Source)
-			}
-
-			return resolver.LocalGetMeta(nil)
-		},
-	}
-
-	return singleClassPropertyFieldsField, nil
-}
-
-func genSingleLocalMetaClassPropertyFields(class *models.SemanticSchemaClass) (graphql.Fields, error) {
-	singleClassPropertyFields := graphql.Fields{}
-	metaPropertyObj := genLocalMetaPropertyObj(class)
-
-	metaPropertyObjField := &graphql.Field{
-		Description: descriptions.GetMetaMetaPropertyDesc,
-		Type:        metaPropertyObj,
-		Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-			return nil, fmt.Errorf("not supported - class property field")
-		},
-	}
-
-	singleClassPropertyFields["meta"] = metaPropertyObjField
-
-	for _, property := range class.Properties {
-		propertyType, err := schema.GetPropertyDataType(class, property.Name)
-
-		if err != nil {
-			return nil, err
-		}
-
-		convertedDataType, err := handleLocalGetMetaNonObjectPropertyDataTypes(*propertyType, class, property)
-
-		if err != nil {
-			return nil, err
-		}
-
-		if *propertyType == schema.DataTypeCRef {
-			singleClassPropertyFields[strings.Title(property.Name)] = convertedDataType
-		} else {
-			singleClassPropertyFields[property.Name] = convertedDataType
-		}
-	}
-
-	return singleClassPropertyFields, nil
-}
-
-func handleLocalGetMetaNonObjectPropertyDataTypes(dataType schema.DataType, class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) (*graphql.Field, error) {
-	metaClassStringPropertyFields := genLocalMetaClassStringPropertyFields(class, property)
-	metaClassTextPropertyFields := genLocalMetaClassTextPropertyFields(class, property)
-	metaClassIntPropertyFields := genLocalMetaClassIntPropertyFields(class, property)
-	metaClassNumberPropertyFields := genLocalMetaClassNumberPropertyFields(class, property)
-	metaClassBooleanPropertyFields := genLocalMetaClassBooleanPropertyFields(class, property)
-	metaClassDatePropertyFields := genLocalMetaClassDatePropertyFields(class, property)
-	metaClassCRefPropertyFields := genLocalMetaClassCRefPropertyObj(class, property)
-
-	switch dataType {
-
-	case schema.DataTypeString:
-		return &graphql.Field{
-			Description: fmt.Sprintf(`%s"%s"`, descriptions.GetMetaPropertyDesc, property.Name),
-			Type:        metaClassStringPropertyFields,
-		}, nil
-
-	case schema.DataTypeText:
-		return &graphql.Field{
-			Description: fmt.Sprintf(`%s"%s"`, descriptions.GetMetaPropertyDesc, property.Name),
-			Type:        metaClassTextPropertyFields,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		}, nil
-
-	case schema.DataTypeInt:
-		return &graphql.Field{
-			Description: fmt.Sprintf(`%s"%s"`, descriptions.GetMetaPropertyDesc, property.Name),
-			Type:        metaClassIntPropertyFields,
-		}, nil
-
-	case schema.DataTypeNumber:
-		return &graphql.Field{
-			Description: fmt.Sprintf(`%s"%s"`, descriptions.GetMetaPropertyDesc, property.Name),
-			Type:        metaClassNumberPropertyFields,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		}, nil
-
-	case schema.DataTypeBoolean:
-		return &graphql.Field{
-			Description: fmt.Sprintf(`%s"%s"`, descriptions.GetMetaPropertyDesc, property.Name),
-			Type:        metaClassBooleanPropertyFields,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		}, nil
-
-	case schema.DataTypeDate:
-		return &graphql.Field{
-			Description: fmt.Sprintf(`%s"%s"`, descriptions.GetMetaPropertyDesc, property.Name),
-			Type:        metaClassDatePropertyFields,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		}, nil
-
-	case schema.DataTypeCRef:
-		return &graphql.Field{
-			Description: fmt.Sprintf(`%s"%s"`, descriptions.GetMetaPropertyDesc, property.Name),
-			Type:        metaClassCRefPropertyFields,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		}, nil
-
-	default:
-		return nil, fmt.Errorf(schema.ErrorNoSuchDatatype)
-	}
-}
-
-func genLocalMetaClassStringPropertyFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	topOccurrencesFields := genLocalMetaClassStringPropertyTopOccurrencesFields(class, property)
-
-	getMetaPointingFields := graphql.Fields{
-
-		"type": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sType", class.Class),
-			Description: descriptions.GetMetaPropertyTypeDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sCount", class.Class),
-			Description: descriptions.GetMetaPropertyCountDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"topOccurrences": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sTopOccurrences", class.Class),
-			Description: descriptions.GetMetaPropertyTopOccurrencesDesc,
-			Type:        graphql.NewList(topOccurrencesFields),
-			Args: graphql.FieldConfigArgument{
-				"first": &graphql.ArgumentConfig{
-					Description: descriptions.FirstDesc,
-					Type:        graphql.Int,
-				},
-				"after": &graphql.ArgumentConfig{
-					Description: descriptions.AfterDesc,
-					Type:        graphql.Int,
-				},
-			},
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	getMetaStringProperty := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
-		Fields:      getMetaPointingFields,
-		Description: descriptions.GetMetaPropertyObjectDesc,
-	}
-
-	return graphql.NewObject(getMetaStringProperty)
-}
-
-func genLocalMetaClassStringPropertyTopOccurrencesFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	getMetaPointingFields := graphql.Fields{
-
-		"value": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sTopOccurrencesValue", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTopOccurrencesValueDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"occurs": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sTopOccurrencesOccurs", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTopOccurrencesOccursDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	getMetaPointing := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sTopOccurrencesObj", class.Class, property.Name),
-		Fields:      getMetaPointingFields,
-		Description: descriptions.GetMetaPropertyTopOccurrencesDesc,
-	}
-
-	return graphql.NewObject(getMetaPointing)
-}
-
-func genLocalMetaClassTextPropertyFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	topOccurrencesFields := genLocalMetaClassTextPropertyTopOccurrencesFields(class, property)
-
-	getMetaPointingFields := graphql.Fields{
-
-		"type": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sType", class.Class),
-			Description: descriptions.GetMetaPropertyTypeDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sCount", class.Class),
-			Description: descriptions.GetMetaPropertyCountDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"topOccurrences": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sTopOccurrences", class.Class),
-			Description: descriptions.GetMetaPropertyTopOccurrencesDesc,
-			Type:        graphql.NewList(topOccurrencesFields),
-			Args: graphql.FieldConfigArgument{
-				"first": &graphql.ArgumentConfig{
-					Description: descriptions.FirstDesc,
-					Type:        graphql.Int,
-				},
-				"after": &graphql.ArgumentConfig{
-					Description: descriptions.AfterDesc,
-					Type:        graphql.Int,
-				},
-			},
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	getMetaTextProperty := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
-		Fields:      getMetaPointingFields,
-		Description: descriptions.GetMetaPropertyObjectDesc,
-	}
-
-	return graphql.NewObject(getMetaTextProperty)
-}
-
-func genLocalMetaClassTextPropertyTopOccurrencesFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	getMetaPointingFields := graphql.Fields{
-
-		"value": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sTopOccurrencesValue", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTopOccurrencesValueDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"occurs": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sTopOccurrencesOccurs", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTopOccurrencesOccursDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	getMetaPointing := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sTopOccurrencesObj", class.Class, property.Name),
-		Fields:      getMetaPointingFields,
-		Description: descriptions.GetMetaPropertyTopOccurrencesDesc,
-	}
-
-	return graphql.NewObject(getMetaPointing)
-}
-
-func genLocalMetaClassIntPropertyFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	getMetaIntFields := graphql.Fields{
-
-		"sum": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sSum", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertySumDesc,
-			Type:        graphql.Float,
-		},
-
-		"type": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sType", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTypeDesc,
-			Type:        graphql.String,
-		},
-
-		"lowest": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sLowest", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyLowestDesc,
-			Type:        graphql.Float,
-		},
-
-		"highest": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sHighest", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyHighestDesc,
-			Type:        graphql.Float,
-		},
-
-		"average": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sAverage", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyAverageDesc,
-			Type:        graphql.Float,
-		},
-
-		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sCount", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyCountDesc,
-			Type:        graphql.Int,
-		},
-	}
-
-	getMetaIntProperty := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
-		Fields:      getMetaIntFields,
-		Description: descriptions.GetMetaPropertyObjectDesc,
-	}
-
-	return graphql.NewObject(getMetaIntProperty)
-}
-
-func genLocalMetaClassNumberPropertyFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	getMetaNumberFields := graphql.Fields{
-
-		"sum": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sSum", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertySumDesc,
-			Type:        graphql.Float,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"type": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sType", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTypeDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"lowest": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sLowest", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyLowestDesc,
-			Type:        graphql.Float,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"highest": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sHighest", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyHighestDesc,
-			Type:        graphql.Float,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"average": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sAverage", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyAverageDesc,
-			Type:        graphql.Float,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sCount", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyCountDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	getMetaNumberProperty := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
-		Fields:      getMetaNumberFields,
-		Description: descriptions.GetMetaPropertyObjectDesc,
-	}
-
-	return graphql.NewObject(getMetaNumberProperty)
-}
-
-func genLocalMetaClassBooleanPropertyFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	getMetaBooleanFields := graphql.Fields{
-
-		"type": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sType", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTypeDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sCount", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyCountDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"totalTrue": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sTotalTrue", class.Class, property.Name),
-			Description: descriptions.GetMetaClassPropertyTotalTrueDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"percentageTrue": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sPercentageTrue", class.Class, property.Name),
-			Description: descriptions.GetMetaClassPropertyPercentageTrueDesc,
-			Type:        graphql.Float,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"totalFalse": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sTotalFalse", class.Class, property.Name),
-			Description: descriptions.GetMetaClassPropertyTotalFalseDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"percentageFalse": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sPercentageFalse", class.Class, property.Name),
-			Description: descriptions.GetMetaClassPropertyPercentageFalseDesc,
-			Type:        graphql.Float,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	getMetaBooleanProperty := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
-		Fields:      getMetaBooleanFields,
-		Description: descriptions.GetMetaPropertyObjectDesc,
-	}
-
-	return graphql.NewObject(getMetaBooleanProperty)
-}
-
-// a duplicate of the string function, this is a separate function to account for future expansions of functionality
-func genLocalMetaClassDatePropertyFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	topOccurrencesFields := genLocalMetaClassDatePropertyTopOccurrencesFields(class, property)
-
-	getMetaDateFields := graphql.Fields{
-
-		"type": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sType", class.Class),
-			Description: descriptions.GetMetaPropertyTypeDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sCount", class.Class),
-			Description: descriptions.GetMetaPropertyCountDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"topOccurrences": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sTopOccurrences", class.Class),
-			Description: descriptions.GetMetaPropertyTopOccurrencesDesc,
-			Type:        graphql.NewList(topOccurrencesFields),
-			Args: graphql.FieldConfigArgument{
-				"first": &graphql.ArgumentConfig{
-					Description: descriptions.FirstDesc,
-					Type:        graphql.Int,
-				},
-				"after": &graphql.ArgumentConfig{
-					Description: descriptions.AfterDesc,
-					Type:        graphql.Int,
-				},
-			},
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	getMetaDateProperty := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
-		Fields:      getMetaDateFields,
-		Description: descriptions.GetMetaPropertyObjectDesc,
-	}
-
-	return graphql.NewObject(getMetaDateProperty)
-}
-
-func genLocalMetaClassDatePropertyTopOccurrencesFields(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	getMetaMetaPointingFields := graphql.Fields{
-
-		"value": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sTopOccurrencesValue", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTopOccurrencesValueDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"occurs": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sTopOccurrencesOccurs", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTopOccurrencesOccursDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	getMetaMetaPointing := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sTopOccurrencesObj", class.Class, property.Name),
-		Fields:      getMetaMetaPointingFields,
-		Description: descriptions.GetMetaPropertyTopOccurrencesDesc,
-	}
-
-	return graphql.NewObject(getMetaMetaPointing)
-}
-
-func genLocalMetaClassCRefPropertyObj(class *models.SemanticSchemaClass, property *models.SemanticSchemaClassProperty) *graphql.Object {
-	getMetaCRefPropertyFields := graphql.Fields{
-
-		"type": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sType", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyTypeDesc,
-			Type:        graphql.String,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sCount", class.Class, property.Name),
-			Description: descriptions.GetMetaPropertyCountDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-
-		"pointingTo": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sPointingTo", class.Class, property.Name),
-			Description: descriptions.GetMetaClassPropertyPointingToDesc,
-			Type:        graphql.NewList(graphql.String),
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	metaClassCRefPropertyConf := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
-		Fields:      getMetaCRefPropertyFields,
-		Description: descriptions.GetMetaPropertyObjectDesc,
-	}
-
-	return graphql.NewObject(metaClassCRefPropertyConf)
-}
-
-func genLocalMetaPropertyObj(class *models.SemanticSchemaClass) *graphql.Object {
-	getMetaPropertyFields := graphql.Fields{
-
-		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%sMetaCount", class.Class),
-			Description: descriptions.GetMetaClassMetaCountDesc,
-			Type:        graphql.Int,
-			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-				return nil, fmt.Errorf("not supported")
-			},
-		},
-	}
-
-	metaPropertyFields := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%sMetaObj", class.Class),
-		Fields:      getMetaPropertyFields,
-		Description: descriptions.GetMetaClassMetaObjDesc,
-	}
-
-	return graphql.NewObject(metaPropertyFields)
-}
