+++ graphqlapi/local/get_meta/property_bool.go
@@ -1,0 +1,71 @@
+package get_meta
+
+import (
+	"fmt"
+
+	"github.com/creativesoftwarefdn/weaviate/graphqlapi/descriptions"
+	"github.com/creativesoftwarefdn/weaviate/models"
+	"github.com/graphql-go/graphql"
+)
+
+func booleanPropertyFields(class *models.SemanticSchemaClass,
+	property *models.SemanticSchemaClassProperty) *graphql.Object {
+	getMetaBooleanFields := graphql.Fields{
+		"type": &graphql.Field{
+			Name:        fmt.Sprintf("Meta%s%sType", class.Class, property.Name),
+			Description: descriptions.GetMetaPropertyTypeDesc,
+			Type:        graphql.String,
+			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+				return nil, fmt.Errorf("not supported")
+			},
+		},
+		"count": &graphql.Field{
+			Name:        fmt.Sprintf("Meta%s%sCount", class.Class, property.Name),
+			Description: descriptions.GetMetaPropertyCountDesc,
+			Type:        graphql.Int,
+			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+				return nil, fmt.Errorf("not supported")
+			},
+		},
+		"totalTrue": &graphql.Field{
+			Name:        fmt.Sprintf("Meta%s%sTotalTrue", class.Class, property.Name),
+			Description: descriptions.GetMetaClassPropertyTotalTrueDesc,
+			Type:        graphql.Int,
+			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+				return nil, fmt.Errorf("not supported")
+			},
+		},
+		"percentageTrue": &graphql.Field{
+			Name:        fmt.Sprintf("Meta%s%sPercentageTrue", class.Class, property.Name),
+			Description: descriptions.GetMetaClassPropertyPercentageTrueDesc,
+			Type:        graphql.Float,
+			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+				return nil, fmt.Errorf("not supported")
+			},
+		},
+		"totalFalse": &graphql.Field{
+			Name:        fmt.Sprintf("Meta%s%sTotalFalse", class.Class, property.Name),
+			Description: descriptions.GetMetaClassPropertyTotalFalseDesc,
+			Type:        graphql.Int,
+			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+				return nil, fmt.Errorf("not supported")
+			},
+		},
+		"percentageFalse": &graphql.Field{
+			Name:        fmt.Sprintf("Meta%s%sPercentageFalse", class.Class, property.Name),
+			Description: descriptions.GetMetaClassPropertyPercentageFalseDesc,
+			Type:        graphql.Float,
+			Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+				return nil, fmt.Errorf("not supported")
+			},
+		},
+	}
+
+	getMetaBooleanProperty := graphql.ObjectConfig{
+		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
+		Fields:      getMetaBooleanFields,
+		Description: descriptions.GetMetaPropertyObjectDesc,
+	}
+
+	return graphql.NewObject(getMetaBooleanProperty)
+}
